Product search is pretty nicely styled, and responsive (for mobile). Functionality is all setup, but I have to make the way I coded it a bit cleaner. Componenets are sharing events that sends irrelevant data to some. Its fine for now, its just client anyway :p

+++ BlazorElectronics/Shared/Products/Search/ProductSearchFilters.cs
@@ -5,8 +5,8 @@ public sealed class ProductSearchFilters
     public DateTime? MinReleaseDate { get; set; }
     public bool? InStock { get; set; } = false;
     public bool? OnSale { get; set; } = false;
-    public int? MinPrice { get; set; } = null;
-    public int? MaxPrice { get; set; } = null;
+    public decimal? MinPrice { get; set; } = null;
+    public decimal? MaxPrice { get; set; } = null;
     public int? MinRating { get; set; } = null;
     
     public List<int> Vendors { get; set; } = new();
